Added:
load_generic_table
This function is more general than load_input_table and later on, should be substituted.
The NumPy array where will be uploaded the table is set up considering the first row of
the input file. The number of columns in the matrix is fetched splitting the first line
using the separator.