Modified:
load_generic_table
Because PoGo added space in the chr name column, it has developed new function functionality to remove it under the control of the flag parameter 'PoGo'.

filter_peptides
Added the input parameter 'out_file_name' as the name of the output file that calls for self.make_sep_file method.